packs/fear-signal-radar/pyproject.toml
[build-system]
requires = []
build-backend = "setuptools.build_meta"
backend-path = [".venv/lib/python3.12/site-packages"]

[project]
name = "fear-signal-radar"
version = "0.1.0"
description = "Fear Signal Radar pack validation tooling"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<10.0.0",
  "jsonschema>=4.0.0,<5.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "httpx>=0.27.0,<1.0.0",
  "testcontainers[postgres]>=4.8.0,<5.0.0",
]

[tool.setuptools]
packages = ["fsra"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = [
  "tests",
]
